+++ tradingstrategy/analysis/__init__.py
@@ -1,1 +1,9 @@
-"""Backtesting success analysis calculations."""
+"""Backtesting success analysis calculations.
+
+.. warning::
+
+    These modules are legacy and moved to `trade-executor` package.
+    This submodule will be removed.
+
+
+"""