+++ tradingstrategy/testing/__init__.py
@@ -1,0 +1,2 @@
+"""Unit and integration testing helpers.
+"""