+++ tradingstrategy/transport/__init__.py
@@ -1,0 +1,6 @@
+"""Data download functionality.
+
+- Different download methods
+
+- Caching
+"""