+++ tradingstrategy/transport/cache.py
@@ -426,10 +426,12 @@ class CachedHTTPTransport:
             Map of pairs -> their trading data availability
 
         """
+
         params = {
-            "pair_ids": list(pair_ids),
+            "pair_ids": ",".join([str(i) for i in pair_ids]),  # OpenAPI comma delimited array
             "time_bucket":  time_bucket.value,
         }
+
         array = self.get_json_response("trading-pair-data-availability", params=params)
 
         # Make to typed and deseralise